# Conflicts:
#	__test__/deleteDTO.test.ts
#	package.json
#	public/templates/config/pom.xml.hbs
#	public/templates/struct/domain-lite/java/infrastructure/persistence/repositoryimpl-ddd.hbs
#	src/index.ts
#	src/interfaces/types.ts
#	src/modules/controller/deleteController.ts
#	src/modules/controller/generateController.ts
#	src/modules/controller/saveControllerConfig.ts
#	src/modules/model/deleteModel.ts
#	src/modules/permission/checkUsageInController.ts
#	src/modules/permission/checkUsageInModel.ts
#	src/schema/modelConfig.ts
#	src/utils/handlebarsHelpers.ts

+++ src/modules/permission/deletePermission.ts
@@ -1,7 +1,5 @@
 import { PermissionConfig, RenderContext } from "../../interfaces/types";
-import { getPermissionConfigPath } from "../../utils/helpers";
-import { checkUsageInController } from "./checkUsageInController";
-import { checkUsageInModel } from "./checkUsageInModel";
+import { getPermissionConfigPath, loadPermissionConfigs } from "../../utils/helpers";
 
 import fs from 'fs-extra'
 import { saveAllPermissions } from "./savePermissions";
@@ -10,13 +8,19 @@ import { saveAllPermissions } from "./savePermissions";
 export const deletePerm = async (context: RenderContext<PermissionConfig>) => {
 
   const permissionPath = getPermissionConfigPath(context.resourceConfig.name, context.basePath)
+  const permissionsType = await loadPermissionConfigs(context.basePath)
 
   if (await (fs.pathExists(permissionPath))) {
-    await checkUsageInController(context)
-    await checkUsageInModel(context)
-    
-    await fs.rm(permissionPath, { recursive: true });
-    await saveAllPermissions(context.basePath);
+    const perm = permissionsType.find(p => p.name === context.resourceConfig.name)
+    if (perm){
+      if (perm.endpoints.length > 0)
+        throw `The permission '${context.resourceConfig.name}' cannot be deleted because it is currently in use`
+      else {
+        await fs.rm(permissionPath, { recursive: true });
+        await saveAllPermissions(context.basePath);
+      }
+
+    }
   }
 
   else throw `Permission '${context.resourceConfig.name}' does not exist.`
